+++ node.h
@@ -1,11 +1,10 @@
 /**
  * node.h
  * 
- * Data structure and function declarations for a graph node.
- * For use in a weighted graph typically used within a search algorithm.
+ * Data structure and procedure declarations for a weighted graph node.
  * 
  * Author: Richard Gale
- * Version: 15th August, 2022 
+ * Version: 25th August, 2022 
  */
 
 #ifndef NODE_H
@@ -13,6 +12,7 @@
 
 #include <stdio.h>
 #include <stdlib.h>
+#include <stdint.h>
 #include <math.h>
 
 /**
@@ -23,7 +23,7 @@ typedef struct node_data* node;
 /**
  * Initialises the node at the provided reference.
  */
-void node_init(node* n_ref, int x, int y, int z, int weight);
+void node_init(node* n_ref, uint16_t x, uint16_t y, uint16_t z, uint8_t weight);
 
 /**
  * Frees the memory allocated to the node at the provided reference.
@@ -39,33 +39,33 @@ node* node_get_came_from(node n);
 /**
  * Returns the x coordinate of the provided node.
  */
-int node_get_x(node n);
+uint16_t node_get_x(node n);
 
 /**
  * Returns the y coordinate of the provided node.
  */
-int node_get_y(node n);
+uint16_t node_get_y(node n);
 
 /**
  * Returns the z coordinate of the provided node.
  */
-int node_get_z(node n);
+uint16_t node_get_z(node n);
 
 /**
  * Returns the estimated total cost of a path if
  * the path goes through the provided node.
  */
-int node_get_f(node n);
+uint32_t node_get_f(node n);
 
 /**
  * Returns the cost of a path from its start to the provided node.
  */
-int node_get_g(node n);
+uint32_t node_get_g(node n);
 
 /**
  * Returns the weight of the provided node.
  */
-int node_get_w(node n);
+uint8_t node_get_w(node n);
 
 /**
  * Sets the node at the provided reference as being the node 
@@ -77,13 +77,13 @@ void node_set_came_from(node* n, node* came_from);
  * Sets the estimated total cost of a path if the path were to go 
  * through the node at the provided reference.
  */
-void node_set_f(node* n, int f);
+void node_set_f(node* n, uint32_t f);
 
 /**
  * Sets the cost of a path from the path's starting point to
  * the node at the provided reference.
  */
-void node_set_g(node* n, int g);
+void node_set_g(node* n, uint32_t g);
 
 /**
  * Prints the node.
